+++ apps/livreGUI/Controller.h
@@ -23,18 +23,15 @@
 #ifndef _Controller_h_
 #define _Controller_h_
 
-#include <livre/core/render/TransferFunction1D.h>
 #include <zeq/types.h>
 
 namespace livre
 {
 
 /**
- * Allows having a set of publishers on different URIs, as well as subscribers
- * that can register different functions for different event types.
- *
- * Controls communication between the GUI and the rendering application.
- * TODO: The widgets for connection can be provided by this interface.
+ * Provides continuous receive on the subscriber for registered events and posts
+ * requests for registered events on the heartbeat event of Livre. This allows
+ * for initializing the GUI once the first livre instance is connected.
  */
 class Controller
 {
@@ -43,46 +40,28 @@ public:
     ~Controller();
 
     /**
-     * TODO: use zeq::connection::Broker to connect livre subscriber to GUI
-     * @param hostname the hostname to connect to
-     * @param port the port of the host to connect to
-     * @return True if the connection was successful, false otherwise
-     */
-    bool connect( const std::string& hostname,
-                  uint16_t port );
-
-    /**
-     * @param uri the publisher URI. If there is no publisher created on the
-     * given URI then add a new one and use it to publish the specified event.
-     * If a publisher object already exists, use it
      * @param event the serialized event to publish
      * @return true if the publication of the event was successful, false
-     * otherwise
+     *         otherwise
      */
-    bool publish( const servus::URI& uri,
-                  const zeq::Event& event );
+    bool publish( const zeq::Event& event );
 
     /**
-     * @param uri the subscriber URI. If there is no subscriber created on the
-     * given URI then add a new one use it to register a function to the
-     * specified event. If a subscriber object already exists, use it.
+     * Register the handler for the given event and publish requests for that
+     * event as well.
+     *
      * @param event the event type of interest
      * @param func the callback function on receive of event
      * @return true if the function was successfully registered
      */
-    bool registerHandler( const servus::URI& uri,
-                          const servus::uint128_t& event,
+    bool registerHandler( const zeq::uint128_t& event,
                           const zeq::EventFunc& func );
 
     /**
-     * @param uri the subscriber URI
      * @param event the event type of interest
-     * @param func the callback function that receives the event
      * @return true if function was successfully deregistered
      */
-    bool deregisterHandler( const servus::URI& uri,
-                            const servus::uint128_t& event,
-                            const zeq::EventFunc& func );
+    bool deregisterHandler( const zeq::uint128_t& event );
 
 private:
     class Impl;